converter/requirements.txt
langchain-google-genai==2.0.10
streamlit==1.46.0
langchain-community==0.3.26
python-dotenv==1.1.1
langchain==0.3.26
langchain-text-splitters==0.3.8
pymupdf==1.26.1
